Update README.md
The goal of my analysis is to answer to these questions:
- Which are the Countries that have won the most and have they changed over the years?
- Can we find correlation between the victories and the GDP per capita? And between the victories and the population?
- Who are the athletes who have won the most?
- How has female participation in games evolved over the years?
- Are men's and women's victories balanced in the nations that have won the most?
To realize my analysis, I used two different datasets obtained from Kaggle. One dataset, Olympic Sport and Medal, 1896-2014, contains more than 35,000 rows and 22 columns. The other dataset, Summer Olympics Medals (1976-2008), contains 15,000 rows and 11 columns including basically the same information - for a shorter period of time and excluding winter games-but it was more complete on information regarding the codes and names of countries dissolved or unified over the years.